move cancelled script and tests to archive
